Language.C.Types

Description

Views of C datatypes. While Language.C.Types.Parse defines datatypes for representing the concrete syntax tree of C types, this module provides friendlier views of C types, by turning them into a data type matching more closely how we read and think about types, both in Haskell and in C. To appreciate the difference, look at the difference between ParameterDeclaration and ParameterDeclaration.

As a bonus, routines are provided for describing types in natural language (English) -- see describeParameterDeclaration and describeType.

Types

data CIdentifier

A type for C identifiers.

unCIdentifier :: CIdentifier -> String

cIdentifierFromString :: Bool -> String -> Either String CIdentifier

data StorageClassSpecifier

Constructors

TYPEDEF

EXTERN

STATIC

AUTO

REGISTER

data TypeQualifier

Constructors

CONST

RESTRICT

VOLATILE

data FunctionSpecifier

Constructors

INLINE

data ArrayType i

Constructors

VariablySized

Unsized

SizedByInteger Integer

SizedByIdentifier i

data Specifiers

Constructors

Specifiers

Fields

storageClassSpecifiers :: [StorageClassSpecifier]

typeQualifiers :: [TypeQualifier]

functionSpecifiers :: [FunctionSpecifier]

data Type i

Constructors

TypeSpecifier Specifiers TypeSpecifier

Ptr [TypeQualifier] (Type i)

Array (ArrayType i) (Type i)

Proto (Type i) [ParameterDeclaration i]

data TypeSpecifier

Constructors

Void

Bool

Char (Maybe Sign)

Short Sign

Int Sign

Long Sign

LLong Sign

Float

Double

LDouble

TypeName CIdentifier

Struct CIdentifier

Enum CIdentifier

Template CIdentifier [TypeSpecifier]

TemplateConst String

TemplatePointer TypeSpecifier

data Sign

Constructors

Signed

Unsigned

data ParameterDeclaration i

Constructors

ParameterDeclaration

Fields

parameterDeclarationId :: Maybe i

parameterDeclarationType :: Type i

Parsing

type TypeNames = HashSet CIdentifier

A collection of named types (typedefs)

type CParser i (m :: Type -> Type) = (Monad m, Functor m, Applicative m, MonadPlus m, Parsing m, CharParsing m, TokenParsing m, LookAheadParsing m, MonadReader (CParserContext i) m, MonadFail m, Hashable i)

All the parsing is done using the type classes provided by the parsers package. You can use the parsing routines with any of the parsers that implement the classes, such as parsec or trifecta.

We parametrize the parsing by the type of the variable identifiers, i. We do so because we use this parser to implement anti-quoters referring to Haskell variables, and thus we need to parse Haskell identifiers in certain positions.

data CParserContext i

cCParserContext :: Bool -> TypeNames -> CParserContext CIdentifier

runCParser

Arguments

:: Stream s Identity Char

=> CParserContext i

-> String
Source name.

-> s
String to parse.

-> ReaderT (CParserContext i) (Parsec s ()) a
Parser. Anything with type forall m. CParser i m => m a is a valid argument.

-> Either ParseError a

Runs a CParser using parsec.

quickCParser

Arguments

:: CParserContext i

-> String
String to parse.

-> ReaderT (CParserContext i) (Parsec String ()) a
Parser. Anything with type forall m. CParser i m => m a is a valid argument.

-> a

Useful for quick testing. Uses "quickCParser" as source name, and throws an error if parsing fails.

quickCParser_

Arguments

:: Bool

-> String
String to parse.

-> ReaderT (CParserContext CIdentifier) (Parsec String ()) a
Parser. Anything with type forall m. CParser i m => m a is a valid argument.

-> a

Like quickCParser, but uses cCParserContext (const False) as CParserContext.

parseParameterDeclaration :: (CParser i m, Pretty i) => m (ParameterDeclaration i)

parseParameterList :: (CParser i m, Pretty i) => m [ParameterDeclaration i]

parseIdentifier :: CParser i m => m i

parseEnableCpp :: CParser i m => m Bool

parseType :: (CParser i m, Pretty i) => m (Type i)

Convert to and from high-level views

data UntangleErr

Constructors

MultipleDataTypes [DeclarationSpecifier]

NoDataTypes [DeclarationSpecifier]

IllegalSpecifiers String [TypeSpecifier]

untangleParameterDeclaration :: ParameterDeclaration i -> Either UntangleErr (ParameterDeclaration i)

tangleParameterDeclaration :: ParameterDeclaration i -> ParameterDeclaration i

To english

describeParameterDeclaration :: Pretty i => ParameterDeclaration i -> Doc ann

describeType :: Pretty i => Type i -> Doc ann