Language.C.Types.Parse

Description

A parser for C99 declarations. Currently, the parser has the following limitations:

Array sizes can only be *, n (where n is a positive integer), x (where x is a C identifier). In C99 they can be arbitrary expressions. See the ArrayType data type.

_Bool, _Complex, and _Imaginary are not present.

Untyped parameter lists (pre-K&R C) are not allowed.

The parser is incremental and generic (see CParser). Pretty and Arbitrary instances are provided for all the data types.

The entry point if you want to parse C declarations is parameter_declaration.

Parser configuration

type TypeNames = HashSet CIdentifier

A collection of named types (typedefs)

data CParserContext i

Constructors

CParserContext

Fields

cpcIdentName :: String

cpcTypeNames :: TypeNames

Function used to determine whether an identifier is a type name.

cpcParseIdent :: forall (m :: Type -> Type). CParser i m => m i

Parses an identifier, *without consuming whitespace afterwards*.

cpcIdentToString :: i -> String

cpcEnableCpp :: Bool

Default configuration

data CIdentifier

A type for C identifiers.

unCIdentifier :: CIdentifier -> String

cIdentifierFromString :: Bool -> String -> Either String CIdentifier

cCParserContext :: Bool -> TypeNames -> CParserContext CIdentifier

Parser type

type CParser i (m :: Type -> Type) = (Monad m, Functor m, Applicative m, MonadPlus m, Parsing m, CharParsing m, TokenParsing m, LookAheadParsing m, MonadReader (CParserContext i) m, MonadFail m, Hashable i)

All the parsing is done using the type classes provided by the parsers package. You can use the parsing routines with any of the parsers that implement the classes, such as parsec or trifecta.

We parametrize the parsing by the type of the variable identifiers, i. We do so because we use this parser to implement anti-quoters referring to Haskell variables, and thus we need to parse Haskell identifiers in certain positions.

runCParser

Arguments

:: Stream s Identity Char

=> CParserContext i

-> String
Source name.

-> s
String to parse.

-> ReaderT (CParserContext i) (Parsec s ()) a
Parser. Anything with type forall m. CParser i m => m a is a valid argument.

-> Either ParseError a

Runs a CParser using parsec.

quickCParser

Arguments

:: CParserContext i

-> String
String to parse.

-> ReaderT (CParserContext i) (Parsec String ()) a
Parser. Anything with type forall m. CParser i m => m a is a valid argument.

-> a

Useful for quick testing. Uses "quickCParser" as source name, and throws an error if parsing fails.

quickCParser_

Arguments

:: Bool

-> String
String to parse.

-> ReaderT (CParserContext CIdentifier) (Parsec String ()) a
Parser. Anything with type forall m. CParser i m => m a is a valid argument.

-> a

Like quickCParser, but uses cCParserContext (const False) as CParserContext.

Types and parsing

identifier_no_lex :: CParser i m => m i

data DeclarationSpecifier

Constructors

StorageClassSpecifier StorageClassSpecifier

TypeSpecifier TypeSpecifier

TypeQualifier TypeQualifier

FunctionSpecifier FunctionSpecifier

declaration_specifiers :: CParser i m => m [DeclarationSpecifier]

data StorageClassSpecifier

Constructors

TYPEDEF

EXTERN

STATIC

AUTO

REGISTER

storage_class_specifier :: CParser i m => m StorageClassSpecifier

data TypeSpecifier

Constructors

VOID

BOOL

CHAR

SHORT

INT

LONG

FLOAT

DOUBLE

SIGNED

UNSIGNED

Struct CIdentifier

Enum CIdentifier

TypeName CIdentifier

Template CIdentifier [[TypeSpecifier]]

TemplateConst String

TemplatePointer TypeSpecifier

type_specifier :: CParser i m => m TypeSpecifier

data TypeQualifier

Constructors

CONST

RESTRICT

VOLATILE

type_qualifier :: CParser i m => m TypeQualifier

data FunctionSpecifier

Constructors

INLINE

function_specifier :: CParser i m => m FunctionSpecifier

data Declarator i

Constructors

Declarator

Fields

declaratorPointers :: [Pointer]

declaratorDirect :: DirectDeclarator i

declarator :: CParser i m => m (Declarator i)

data DirectDeclarator i

Constructors

DeclaratorRoot i

ArrayOrProto (DirectDeclarator i) (ArrayOrProto i)

DeclaratorParens (Declarator i)

direct_declarator :: CParser i m => m (DirectDeclarator i)

data ArrayOrProto i

Constructors

Array (ArrayType i)

Proto [ParameterDeclaration i]

array_or_proto :: CParser i m => m (ArrayOrProto i)

data ArrayType i

Constructors

VariablySized

Unsized

SizedByInteger Integer

SizedByIdentifier i

array_type :: CParser i m => m (ArrayType i)

data Pointer

Constructors

Pointer [TypeQualifier]

pointer :: CParser i m => m Pointer

data ParameterDeclaration i

Constructors

ParameterDeclaration

Fields

parameterDeclarationSpecifiers :: [DeclarationSpecifier]

parameterDeclarationDeclarator :: DeclaratorOrAbstractDeclarator i

data DeclaratorOrAbstractDeclarator i

Constructors

IsDeclarator (Declarator i)

IsAbstractDeclarator (AbstractDeclarator i)

parameter_declaration :: CParser i m => m (ParameterDeclaration i)

parameter_list :: CParser i m => m [ParameterDeclaration i]

data AbstractDeclarator i

Constructors

AbstractDeclarator

Fields

abstractDeclaratorPointers :: [Pointer]

abstractDeclaratorDirect :: Maybe (DirectAbstractDeclarator i)

abstract_declarator :: CParser i m => m (AbstractDeclarator i)

data DirectAbstractDeclarator i

Constructors

ArrayOrProtoHere (ArrayOrProto i)

ArrayOrProtoThere (DirectAbstractDeclarator i) (ArrayOrProto i)

AbstractDeclaratorParens (AbstractDeclarator i)

direct_abstract_declarator :: CParser i m => m (DirectAbstractDeclarator i)

YACC grammar

The parser above is derived from a modification of the YACC grammar for C99 found at http://www.quut.com/c/ANSI-C-grammar-y-1999.html, reproduced below.

%token IDENTIFIER TYPE_NAME INTEGER

%token TYPEDEF EXTERN STATIC AUTO REGISTER INLINE RESTRICT
%token CHAR SHORT INT LONG SIGNED UNSIGNED FLOAT DOUBLE CONST VOLATILE VOID
%token BOOL COMPLEX IMAGINARY
%token STRUCT UNION ENUM

%start parameter_list
%%

declaration_specifiers
	: storage_class_specifier
	| storage_class_specifier declaration_specifiers
	| type_specifier
	| type_specifier declaration_specifiers
	| type_qualifier
	| type_qualifier declaration_specifiers
	| function_specifier
	| function_specifier declaration_specifiers
	;

storage_class_specifier
	: TYPEDEF
	| EXTERN
	| STATIC
	| AUTO
	| REGISTER
	;

type_specifier
	: VOID
	| CHAR
	| SHORT
	| INT
	| LONG
	| FLOAT
	| DOUBLE
	| SIGNED
	| UNSIGNED
	| BOOL
	| COMPLEX
	| IMAGINARY
 	| STRUCT IDENTIFIER
	| UNION IDENTIFIER
	| ENUM IDENTIFIER
	| TYPE_NAME
	;

type_qualifier
	: CONST
	| RESTRICT
	| VOLATILE
	;

function_specifier
	: INLINE
	;

declarator
	: pointer direct_declarator
	| direct_declarator
	;

direct_declarator
	: IDENTIFIER
	| '(' declarator ')'
	| direct_declarator '[' type_qualifier_list ']'
	| direct_declarator '[' type_qualifier_list * ']'
	| direct_declarator '[' * ']'
 	| direct_declarator '[' IDENTIFIER ']'
	| direct_declarator '[' INTEGER ']'
	| direct_declarator '[' ']'
	| direct_declarator '(' parameter_list ')'
	| direct_declarator '(' ')'
	;

pointer
	: *
	| * type_qualifier_list
	| * pointer
	| * type_qualifier_list pointer
	;

type_qualifier_list
	: type_qualifier
	| type_qualifier_list type_qualifier
	;

parameter_list
	: parameter_declaration
	| parameter_list ',' parameter_declaration
	;

parameter_declaration
	: declaration_specifiers declarator
	| declaration_specifiers abstract_declarator
	| declaration_specifiers
	;

abstract_declarator
	: pointer
	| direct_abstract_declarator
	| pointer direct_abstract_declarator
	;

direct_abstract_declarator
	: '(' abstract_declarator ')'
	| '[' ']'
	| direct_abstract_declarator '[' ']'
	| '[' * ']'
	| direct_abstract_declarator '[' * ']'
	| '[' IDENTIFIER ']'
	| direct_abstract_declarator '[' IDENTIFIER ']'
	| '[' INTEGER ']'
	| direct_abstract_declarator '[' INTEGER ']'
	| '(' ')'
	| '(' parameter_list ')'
	| direct_abstract_declarator '(' ')'
	| direct_abstract_declarator '(' parameter_list ')'
	;

%%
#include <stdio.h>

extern char yytext[];
extern int column;

void yyerror(char const *s)
{
	fflush(stdout);
	printf("n%*sn%*sn", column, "^", column, s);
}

Testing utilities

cIdentStart :: [Char]

cIdentLetter :: [Char]

cReservedWords :: HashSet String

isTypeName :: Bool -> TypeNames -> String -> Bool