Language.C.Inline.Interruptible

Description

interruptible variants of the Language.C.Inline quasi-quoters, to call interruptible C code. See https://downloads.haskell.org/~ghc/latest/docs/html/users_guide/ffi.html#ffi-interruptible for more information.

This module is intended to be imported qualified:

import qualified Language.C.Inline.Interruptible as CI

Documentation

exp :: QuasiQuoter

C expressions.

pure :: QuasiQuoter

Variant of exp, for use with expressions known to have no side effects.

BEWARE: use this function with caution, only when you know what you are doing. If an expression does in fact have side-effects, then indiscriminate use of pure may endanger referential transparency, and in principle even type safety.

block :: QuasiQuoter

C code blocks (i.e. statements).